+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/InjectedClassTransformer.kt
@@ -704,7 +704,7 @@ internal class InjectedClassTransformer(
       factoryClass = factoryCls,
       sourceTypeParameters = targetClass,
       returnTypeProvider = { typeParams ->
-        metroSymbols.metroFactory.typeWith(targetClass.symbol.typeWithParameters(typeParams))
+        factoryCls.symbol.typeWithParameters(typeParams)
       },
       targetConstructor = factoryConstructor,
       parameters = dedupedMerged,